extract data mapping
