+++ src/app/[locale]/viewer/_components/submodel-elements/document-component/DocumentComponent.tsx
@@ -1,10 +1,6 @@
 import { InfoOutlined, InsertDriveFileOutlined, OpenInNew } from '@mui/icons-material';
 import { Box, Button, IconButton, Link, styled, Typography } from '@mui/material';
 import {
-    File,
-    ISubmodelElement,
-    MultiLanguageProperty,
-    Property,
     SubmodelElementCollection,
 } from '@aas-core-works/aas-core3.0-typescript/types';
 import { DataRow } from 'components/basics/DataRow';
@@ -12,36 +8,24 @@ import { PdfDocumentIcon } from 'components/custom-icons/PdfDocumentIcon';
 import { useState } from 'react';
 import {
     findAllSubmodelElementsBySemanticIdsOrIdShort,
-    findSubmodelElementBySemanticIdsOrIdShort,
-    getTranslationText,
 } from 'lib/util/SubmodelResolverUtil';
 import { DocumentDetailsDialog } from './DocumentDetailsDialog';
-import { isValidUrl } from 'lib/util/UrlUtil';
 import { useAsyncEffect } from 'lib/hooks/UseAsyncEffect';
-import { useAasOriginSourceState } from 'components/contexts/CurrentAasContext';
-import { encodeBase64 } from 'lib/util/Base64Util';
 import { checkFileExists } from 'lib/services/search-actions/searchActions';
-import { useLocale, useTranslations } from 'next-intl';
-import { findIdShortForLatestElement } from './DocumentUtils';
+import { useTranslations } from 'next-intl';
+import { DocumentClassification } from 'app/[locale]/viewer/_components/submodel-elements/document-component/DocumentClassification';
+import {
+    useFileViewObject
+} from 'app/[locale]/viewer/_components/submodel-elements/document-component/useDocumentVersionData';
 import {
     DocumentSpecificSemanticId,
     DocumentSpecificSemanticIdIrdi,
-    DocumentSpecificSemanticIdIrdiV2,
-} from './DocumentSemanticIds';
-import { DocumentClassification } from 'app/[locale]/viewer/_components/submodel-elements/document-component/DocumentClassification';
+} from 'app/[locale]/viewer/_components/submodel-elements/document-component/DocumentSemanticIds';
 
 type DocumentComponentProps = {
     readonly submodelElement: SubmodelElementCollection;
     readonly hasDivider: boolean;
     readonly submodelId: string;
-};
-
-export type FileViewObject = {
-    mimeType: string;
-    title: string;
-    digitalFileUrl: string;
-    previewImgUrl: string;
-    organizationName: string;
 };
 
 const StyledImageWrapper = styled(Box)(({ theme }) => ({
@@ -66,12 +50,11 @@ const StyledImageWrapper = styled(Box)(({ theme }) => ({
 
 export function DocumentComponent(props: DocumentComponentProps) {
     const t = useTranslations('components.documentComponent');
-    const locale = useLocale();
-    const [fileViewObject, setFileViewObject] = useState<FileViewObject>();
     const [detailsModalOpen, setDetailsModalOpen] = useState(false);
-    const [aasOriginUrl] = useAasOriginSourceState();
     const [imageError, setImageError] = useState(false);
     const [fileExists, setFileExists] = useState(true);
+    const fileViewObject = useFileViewObject(props.submodelElement, props.submodelId);
+
 
     useAsyncEffect(async () => {
         if (fileViewObject?.digitalFileUrl) {
@@ -79,10 +62,6 @@ export function DocumentComponent(props: DocumentComponentProps) {
             setFileExists(checkResponse.isSuccess && checkResponse.result);
         }
     }, [fileViewObject?.digitalFileUrl]);
-
-    useAsyncEffect(async () => {
-        setFileViewObject(getFileViewObject());
-    }, [props.submodelElement]);
 
     const handleDetailsClick = () => {
         setDetailsModalOpen(true);
@@ -115,144 +94,6 @@ export function DocumentComponent(props: DocumentComponentProps) {
             )}
         </StyledImageWrapper>
     );
-
-    function findIdShortForLatestDocument(submodelElement: SubmodelElementCollection) {
-        return findIdShortForLatestElement(
-            submodelElement,
-            'DigitalFile',
-            DocumentSpecificSemanticId.DigitalFile,
-            DocumentSpecificSemanticIdIrdi.DigitalFile,
-            DocumentSpecificSemanticIdIrdiV2.DigitalFile,
-        );
-    }
-
-    function findIdShortForLatestPreviewImage(submodelElement: SubmodelElementCollection) {
-        return findIdShortForLatestElement(
-            submodelElement,
-            'PreviewFile',
-            DocumentSpecificSemanticId.PreviewFile,
-            DocumentSpecificSemanticIdIrdi.PreviewFile,
-            DocumentSpecificSemanticIdIrdiV2.PreviewFile,
-        );
-    }
-
-    function getDigitalFile(versionSubmodelEl: ISubmodelElement, submodelElement: ISubmodelElement) {
-        const digitalFile = {
-            digitalFileUrl: '',
-            mimeType: '',
-        };
-
-        if (isValidUrl((versionSubmodelEl as File).value)) {
-            digitalFile.digitalFileUrl = (versionSubmodelEl as File).value || '';
-            digitalFile.mimeType = (versionSubmodelEl as File).contentType;
-        } else if (props.submodelId && submodelElement.idShort && props.submodelElement?.idShort) {
-            const submodelElementPath =
-                props.submodelElement.idShort +
-                '.' +
-                submodelElement.idShort +
-                '.' +
-                findIdShortForLatestDocument(submodelElement as SubmodelElementCollection);
-
-            digitalFile.digitalFileUrl =
-                aasOriginUrl +
-                '/submodels/' +
-                encodeBase64(props.submodelId) +
-                '/submodel-elements/' +
-                submodelElementPath +
-                '/attachment';
-
-            digitalFile.mimeType = (versionSubmodelEl as File).contentType;
-        }
-
-        return digitalFile;
-    }
-
-    function getPreviewImageUrl(versionSubmodelEl: ISubmodelElement, submodelElement: ISubmodelElement) {
-        let previewImgUrl;
-
-        if (isValidUrl((versionSubmodelEl as File).value)) {
-            previewImgUrl = (versionSubmodelEl as File).value ?? '';
-        } else if (props.submodelId && submodelElement.idShort && props.submodelElement?.idShort) {
-            const submodelElementPath =
-                props.submodelElement.idShort +
-                '.' +
-                submodelElement.idShort +
-                '.' +
-                findIdShortForLatestPreviewImage(submodelElement as SubmodelElementCollection);
-
-            previewImgUrl =
-                aasOriginUrl +
-                '/submodels/' +
-                encodeBase64(props.submodelId) +
-                '/submodel-elements/' +
-                submodelElementPath +
-                '/attachment';
-        }
-
-        return previewImgUrl ?? '';
-    }
-
-    function extractDocumentVersionData(documentVersion: SubmodelElementCollection, fileViewObject: FileViewObject) {
-        const title = findSubmodelElementBySemanticIdsOrIdShort(documentVersion.value, null, [
-            DocumentSpecificSemanticId.Title,
-            DocumentSpecificSemanticIdIrdi.Title,
-            DocumentSpecificSemanticIdIrdiV2.Title,
-        ]);
-        fileViewObject.title = getTranslationText(title as MultiLanguageProperty, locale);
-
-        const file = findSubmodelElementBySemanticIdsOrIdShort(documentVersion.value, 'DigitalFile', [
-            DocumentSpecificSemanticId.DigitalFile,
-            DocumentSpecificSemanticIdIrdi.DigitalFile,
-            DocumentSpecificSemanticIdIrdiV2.DigitalFile,
-        ]);
-        fileViewObject = file
-            ? {
-                  ...fileViewObject,
-                  ...getDigitalFile(file, documentVersion),
-              }
-            : fileViewObject;
-
-        const preview = findSubmodelElementBySemanticIdsOrIdShort(documentVersion.value, 'PreviewFile', [
-            DocumentSpecificSemanticId.PreviewFile,
-            DocumentSpecificSemanticIdIrdi.PreviewFile,
-            DocumentSpecificSemanticIdIrdiV2.PreviewFile,
-        ]);
-        fileViewObject.previewImgUrl = preview ? getPreviewImageUrl(preview, documentVersion) : '';
-
-        const organization = findSubmodelElementBySemanticIdsOrIdShort(documentVersion.value, 'OrganizationName', [
-            DocumentSpecificSemanticId.OrganizationName,
-            DocumentSpecificSemanticIdIrdi.OrganizationName,
-            DocumentSpecificSemanticIdIrdiV2.OrganizationShortName,
-        ]);
-        fileViewObject.organizationName = (organization as Property).value || '';
-
-        return fileViewObject;
-    }
-
-    function getFileViewObject(): FileViewObject {
-        let fileViewObject: FileViewObject = {
-            mimeType: '',
-            title: props.submodelElement?.idShort ?? '',
-            organizationName: '',
-            digitalFileUrl: '',
-            previewImgUrl: '',
-        };
-        if (!props.submodelElement?.value) return fileViewObject;
-
-        const documentVersion = findSubmodelElementBySemanticIdsOrIdShort(
-            props.submodelElement.value,
-            'DocumentVersion',
-            [
-                DocumentSpecificSemanticId.DocumentVersion,
-                DocumentSpecificSemanticIdIrdi.DocumentVersion,
-                DocumentSpecificSemanticIdIrdiV2.DocumentVersion,
-            ],
-        ) as SubmodelElementCollection;
-        if (documentVersion?.value) {
-            fileViewObject = extractDocumentVersionData(documentVersion, fileViewObject);
-        }
-        return fileViewObject;
-    }
 
     function getDocumentClassificationCollection() {
         return findAllSubmodelElementsBySemanticIdsOrIdShort(props.submodelElement.value, 'DocumentClassification', [
